FIX broken JSPUI compilation.
I had to downgrade to Tomcat 7 as Tomcat 8 in Ubuntu as the there was a
known bug preventing me from running dspace:
https://bugs.launchpad.net/ubuntu/+source/tomcat8/+bug/1606331

This means downgrading compatibility to Java 7 which caused some
compilation issues.

Also, looks like statspace.showcase.handle actually takes a collection.
I had it set to a boolean and it causes JSPUI to die. Changed it to
empty like how it was back when it was in build.properties.