ADD new page for viewing submitted content.
Partial work on #10, with a lot of placeholders.

All new custom pages and packages are placed in a new ubc directory or
namespace, to reduce confusion about what's original code and what's
changed.

There were some radical changes required for the item display change, so
I copied the original item display page and modified it to match. A lot
of the HTML was being generated in the custom tag ItemTag.java. It made
it really annoying to modify so I've moved the logic into the new
MetadataRetriever and BitstreamRetriever objects. The html is now in the
JSP, where it belongs.

Appearantly, there is a thumbnail generation feature built in that
requires ImageMagick to be installed on the server. The thumbnail
generation has to be executed externally, e.g.: in the dspace
installation directory, run `./bin/dspace filter-media`

I've enabled this in the configuration file, still need to adjust the
deployment scripts to reflect the new changes though.