ADD combine class for metadata&bitstream retrieval
ItemRetriever combines BitstreamRetriever and MetadataRetriever. There's
a lot of logic that shouldn't be in JSPs or in custom tags, the
retrievers' copied those logic for uniform and easier access. It was
still a bit too much scriptlets in a JSP for my taste though, so I made
a unified class where all non-JSP logic can go.

BitstreamRetriever renamed to ItemBitstreamRetriever.
MetadataRetriever renamed to ItemMetadataRetriever.

Metadata retriever can now deal with multi-value fields such as
subjects/tags.