EDIT search results looks & organization.
Search result boxes now are shared height. The 'view' button is now
aligned to the bottom of the box. The thumbnail will try its best to
proportionally scale the images for a uniform look. A placeholder icon
will be shown for entries that don't have a thumbnail yet.

Note, using CSS3 flexbox features, might be a bit buggy on IE.