ADD skip select collection step.
Since we only have one collection, it seems a bit useless to have a
Select Collection step. Unfortunately, the Select Collection step is one
that's unskippable by configuration, since DSpace figured that everyone
has to have that step.

This seems a bit complicated and time consuming to implement on the
servlet side of things. Seems like we need to insert a custom filter
into the submissions servlet that overrides the HttpServletRequest
object with a wrapper to include the collection parameter that we want
to send. This is intentional as the API made a deliberate decision to
not allow easy HttpServletRequest modifications.

I've made a less ideal workaround on the front end instead. It'll still
go to the Select Collection page but I've added a script to that page
that'll automatically select a collection and submit the form. For the
few seconds that this page is visible, it'll be hidden under an overlay
with a please wait message.

Implements #20