ADD new subject selection for submitting items.
Using the new subject tree, there is now a custom tailored UI for
selecting subjects. The original selection form is still being used,
just hidden. Another set of more friendly UI is used to select subjects.
Changes are mirroed to the hidden selection form, which is still being
used on submit by the servlets to make the actual change.

edit-metadata.jsp is just a copy of the original with some
modifications. The subjects UI is kinda bolted on to the original.

It's really crappy practice to use so much scriptlets in a JSP, but
that's what DSpace came with...

This finishes implementing #9.