REMOVE statspace schema file.
The statspace schema is problematic as it keeps crashing the
`./bin/dspace create-administrator` command cause it keeps trying to
recreate the statspace schema even though it already exists. I'm not
sure what the original plan for it was as the commit message didn't say
much and it doesn't seem to be used in the code anywhere. I'm removing
it for now out of convenience. Hopefully it won't cause trouble
elsewhere.