FIX broken copyright selection step.
Attempt to fix the broken copyright selection step. I'm not sure if this
will fix it as it was pretty difficult to debug. Weirdly, if you
download the docker container and run it locally, everything works fine.
So something about the verf/prod servers are different somehow.

I was able to reproduce a similar effect when the PluginManager was
given an empty plugin name. As far as I figure, that makes the
PluginManager return null and the JSONRequest just 404s. Now, on the
real server, I'm speculating that the plugin name wasn't actually empty,
but the effect was the same. I think PluginManager couldn't find
CreativeCommonsJSONRequest to instantiate. This seems similar to cases
where Java security settings are preventing the type of Class abuse
that's being used here.

As a workaround, I've manually instantiated the
CreativeCommonsJSONRequest object when it encounters a matching plugin
name. Given the limited knowledge I have of the verf/prod environment,
this was the faster solution for now.