EDIT view item, show subjects, replace placeholder
Not as pretty as I'd like it due to the lack of time. The tags panels
now the subjects panel. A new results class was created to help with
parsing the subject format into its different levels. All
non-implemented elements were commented out. Placeholders have been
replaced with actual values.