EDIT Sign In & Register now available again.
For the nav bar and home page, I just restored commented out parts that
seems to have be prepared ahead of time for the ending of the soft
launch.

For the new user registration page, I mostly restored the original
DSpace registration page as I'm a bit confused by the existing code. The
descriptions say that this is a form for registering your email for
interest, but the thing is, it points to the registration servlet, so
they *should* be registering for an account, not really registering
interest. Don't know what the original plan was, maybe they were
planning to send the emails afterwards using the registration tokens?