ADD Suggested Uses, Tips and Discoveries field.
"Suggested Uses and Tips" and "What We Learned" seems to be similar
enough in intent to each other that I think we should merge them into
one field. I've named this field "Suggested Uses, Tips and Discoveries".
This also works out better metadata wise because it fits the
dcterms.instructionalMethod registry that I was using.

This finishes implementing #17.