EDIT TinyMCE, edit raw html, load only on new item
When I edited an item as an administrator, there was a different
interface of all textareas, and all the textareas had an TinyMCE
instance on it. This added unnecessary HTML tags to all entries. To
prevent this in the future, I've only enabled TinyMCE to load on new
item submission form and enabled the ability to edit the raw html, so at
least we can fix it if anything similar happens again.