ADD automatic periodic thumbnail generation.
DSpace thumbnail generation requires calling an external script, e.g.:
`./bin/dspace filter-media`. They suggest putting it on cron for
automatic periodic thumbnail updates. Unfortunately, this is problematic
for the docker setup we're using. Docker recommends that only one
process is run for each image and is built around that. To run cron,
we'd need to have an additional cron service running alongside. This
extra process won't benefit from benefits such as docker monitoring
either, so if it went down, it'd do so silently.

To get around this, we're using the built in Java scheduling
capabilities through the Quartz Scheduler. This means scheduling is part
of the normal single DSpace process that we run.

MediaFitlerManager is what the dspace script calls for filter-media.
Unfortunately, it was built to be run only in the command line and calls
System.exit() when it finishes. This kills the entire web server. I copy
and pasted their code and made a function where all System.exit() were
replaced with returns. This does mean that it's not very robust right
now, if a thumbnail generation fails, we have little insight into what's
wrong. Will need to revisit later.

Currently set to go off every 5 minutes. Should be extended to a longer
interval when testing is finished. 5 minutes allows for faster feedback
during testing.