EDIT labels for Prereq, Resource Types, Subjects.
Prerequisites, add explanation that it should be entered as one per line
(pressing the + button to get more lines, etc).

Resource Types, add instructions on how to do multiple selections,
(e.g.: holding ctrl on Windows, cmd on Mac). This was in previously, so
I've restored the old wording.

Subjects, add the asterisk indicating a required field.

This implements #22