FIX copyright step being skippable.
Users are now forced to make a selection on the copyright selection page
before they can click on the Next Step button.

This is just a javascript toggle adding/removing the CSS class
"disabled" to the Next Step button. What's interesting is that when you
select "Creative Commons", it places a set of radio buttons onto the
page for additional options. These additional options seem to come with
their own validation script so I didn't have to write one for them too.

Now serving the creative-commons.jsp file from the statspace dir for
this modification.

Fixes #30.