ADD TinyMCE to single line inputs on item submit.
Each entry in Related Materials (and other such plain text input fields
except for the tinymce ones) requires you to hand write html if you want
to add formatting. Given that URLs will be common in Related Materials,
there should be a more user friendly way to input them. The easiest way
to do this was to just add a more stripped down TinyMCE to those fields,
so just going to do that for now. Sadly, this does quite ruin the
performance due to the addition of extra fields being quite slow. Not
sure what a good alternative would be.

This implements #24.