EDIT file upload, hide bitstream & checksum.
UI instruction changes.

Hide the "primary bitstream" selection. I'm not sure what that setting
does anyways. Probably meant to select a primary file for thumbnails?
But it wasn't being used that way in the original dspace code, so not
sure.

Hide all mentions of checksums, since normal users aren't going to know
what it is.