EDIT subject 3rd lvl optional, multi&single select.
Stats wants single subject selections, Bio wants multi subject
selections. I've reworked the subject selection to support both. If the
subjects field defined in `input-forms.xml` has repeatable set to true,
then it'll provide the multiselect version. Otherwise, it'll provide the
single select version.

Stats wants the subject selection to be able to stop at the 2nd level,
with the 3rd level optional. So I've made the default option on the 3rd
level the "None/Other" option to stand in for only 2 levels of
selection.

Note: Due to the 3rd level optional function, the provided subject list
in `input-forms.xml` MUST have the appropriate values, e.g.: If you have
A >>> B >>> C and A >>> D >>> E, you must have A >>> B and A >>> D.