ADD TinyMCE spellchecking, modify "Add" filter lbl
Modified the "Add" filter button to read "Apply". Stats feels "Apply" is
more intuitive.

Configure TinyMCE to let the browser's own spell checker help the user
out.  Spellchecking requested by Stats. Note that TinyMCE has a more
powerful spellchecker but it isn't part of the free community version
we're using, you have to upgrade to a paid subscription for it.