FIX short session timeouts.
Pan found out what was causing the short session timeouts!

DSpace checks to see if the IP you logged in with is different from the
IP you're currently visiting the site with. If the IP has changed, it
won't let you authenticate. This was causing short session timeout
issues due to us using 2 load balancers. If a user hits a different load
balancer it looks as if their IP changed and so their authentication
gets rejected.

Removing the check should fix the issue.

Also note that mobile users will see a lot of IP changes
so removing this will help mobile compatibility too.

Fixes #26.