ADD alt lang field, EDIT search filter order, case
Added a new alternative language field, for linking to different
language versions of this item. This is complicated by the fact that you
can't link to in-progress submissions. So might need curators to go
through and manually link different language submissions together. Or
develop some way of automatically determining two submissions are the
same, just in different languages.

Since the "Equals" search filter is a bit too strict to be useful, I've
demoted it to the second option and promoted the "Contains" filter to
default.

There's some documentation indicating that
`webui.browse.metadata.case-insensitive` should be enabled for case
insensitive search. But reading the comments in the config file, it
seems that it only enables the case-insensitive metadata display. Not
sure which is true, but it seems useful either way, so I've enabled it
just to be sure.