ADD instructors-only permission for viewing items.
Basically lets submitters select whether they want to let students see
the resource they're uploading. Items restricted to instructors only
won't show up to users who aren't instructors.

Users who are system admins and who are curators also have access. The
actual permission check is a simple check to see if the user is a system
admin or if they're in a group that has the word "instructor" or
"curator" in it.

InstructorAccessChecker is intended as the additional item level layer
permission check counterpart to DSpace's AuthorizeManager. Maybe it
should raise AuthorizeExceptions like AuthorizeManager instead of
returning booleans, not sure about the design choice. Was thinking that
if there's a demand for a more complex permissions system, we can
revisit it for refactoring if needed.

Note that because the Solr search engine and the dspace database are
decoupled in dspace, it's a bit difficult to enforce dspace permissions
on the search results. This unfortunately means that users can infer
that the existence of items they don't have access to. E.g.: The
Subjects sidebar in search results, there's a count of how many items
are in each subject. The count includes both items the user has access
to and the ones they don't.

ItemMetadataRetriever was modified to not require PageContext.
PageContext is only needed to retrieve the more user friendly labels for
metadata fields. Turns out I never use the labels, but since it's there
already, I'm keeping it around just in case.