EDIT creative commons link & hints.
When you choose the creative commons license, you are able to customize
it. There used to be an hint icon with additional explanations for what
you're customizing, but since it only shows up when you hover over the
icon, it can be missed. So I've moved it out to be displayed as muted
text with the options. Note that the hint comes from creative commons
and I can't change it.