ADD show date submitted for curator tasks.
Submissions waiting for curator approval should show "date submitted" so
curators have an idea of how far behind they are. This was more
complicated than expected because for some reason, the date submitted
wasn't being recorded automatically already. So I have to make it record
the date submitted. Honestly, I'm not happy with the implementation here
as there's probably a better place to automatically tag on the date
submitted. But it was easy to put it in JSPCompleteStep, so I've put it
there. (Doesn't it make sense to have the submission save code in
*Complete*Step?! Why is it empty?! Probably missing something here...)

Only submissions made after this commit will have submission dates. Used
the dc.date.submitted metadata field.