ADD mark individual files as "instructor only".
Instructors want to be able to have entries that are otherwise open to
the public but have some files that are hidden except from instructors.
E.g.: for an exam, they'd want the actual exam/answer key to be
"instructor only", but students can still access sample questions.

Looks like bitstreams can have the same kind of metadata that submission
items have. So I've added the same dcterms.accessRight that  items have
for storing permissions to bitstreams. Permissions can also be checked
using the same UBCAccessChecker that items use.

This only marks the files "instructor only", it doesn't actually enforce
the policy.

Note that curators CANNOT edit the files. This is built-in dspace
policy, not something I built.

Partial #40.