ADD enforcement for "instructor only" files.
Files that are marked as "instructor only" will not show up to users who
don't have access. Direct links to the file will throw the same error as
if the file doesn't exist.

Discovered that UBCAccessChecker cannot create it's own DSpace Context,
as that context won't contain information about the logged in user. This
necessitated overhauling some of the Retrievers and Servlets to provide
this Context.

Partial #40.