ADD item instructor-only overrides bitstream's.
Setting instructor-only at the item level should override setting
instructor-only on each bitstream. To make this less complicated for
users, we can hide the instructor-only checkboxes for individual files
if the item level instructor-only is set.

If the bitstream instructor-only checkboxes are hidden, there is a note
saying that "item level" instructor-only is on so all files are
considered to be instructor-only.

Note that this means, by default, if the item level instructor-only is
lifted, all items are set to everyone permission

Last commit to implement #40.