FIX submission's date input broken on Safari.
Looks like there was an edge case with Safari where on certain window
sizes, the date input gets shrinked so small that you can't see what
you've typed into it. I've shrinked the size of the month input and
increased the size of the date input as a workaround.

Fixes #35.