ADD basic video/audio HTML5 player.
Using the video.js player. Should work for both video and audio files.

Currently a bit dumb in that it assumes all mp4, webm, ogg files are
playable videos. IE/Edge only supports mp4, so will error out on the
other files. mp4 is also limited to h264 video and aac audio which we
don't check. Will try to improve the file exclusion a bit so that we
don't try to present non-compatible media files for streaming.

A complete solution would include a backend that converts non-streaming
compatible videos into an acceptable format. Sadly, this would be a
major feature that we don't have time for.

To accomodate visual changes due to the streaming player, I've redid the
display item's file listing.