FIX CSS rows extending beyond its parent.
Some elements with the class row also have the container class. Because
the container class is supposed to only be on a top level element, this
causes unexpected behaviours. I guess it wasn't causing issues in the
older versions of Bootstrap, but with the upgrade to latest Bootstrap 3,
it started causing floating elements to disappear.

Most notably, I couldn't add people or groups for access to
collections.