ADD restricted access system for items & files.
This is a more generalized permission system that can also be adapted
for StatSpace. To adapt to another system, the main method to change is
UBCAccessChecker's hasRestrictedAccess(). The difference between
CWSEISpace and StatSpace permission is that StatSpace wants to restrict
to instructors only while CWSEISpace wants to only hide things from
users who only have anonymous access.

Also includes the fix for not being able to upload large files.

And a new fix where editing the file descriptions to empty wouldn't
remove the description as it's supposed to.

Merged from work done in StatSpace:
https://github.com/ubc/DSpace/commit/ce0db97504bc450b99490db288c8477c1defc544
https://github.com/ubc/DSpace/commit/0341c654c3679bfdf35f27c5e23335017eb9b063
https://github.com/ubc/DSpace/commit/3d311cc91e80ef8ab5886b6c8879fd512395960b
https://github.com/ubc/DSpace/commit/c9d8dc23f4054a5586c6db95323173ff7aac06a4
https://github.com/ubc/DSpace/commit/f930f3e1557a99d42853943f44f4e3680d45bb2f
https://github.com/ubc/DSpace/commit/4b1aed78c97ea610e1d98f1460559ddfee175fa7

Implements #49.