EDIT thumbnail/title/search linking,force download
In display-item, changed the file thumbnails and titles to link directly
to the file and let the browser decide how to handle those files. The
download button will force download all files even if the browser knows
how to handle it.

On the search result page, make the item title a link to the item.
Removed the "View" button as redundant, since the thumbnail and the
title are both links to the item already.

Part of #48.