EDIT make resource type pills into links.
The resource type is listed on the search page and the display item
page. It's the green bootstrap pill elements. The link takes the user to
a search result page listing only that resource type.

Had to add a max size to the resource type pill cause it was extruding
beyond the boundary of the result boxes.