EDIT Overhaul display item, try make more compact.
Try to make it more space efficient. Also ended up splitting it into
several different components, hoping that it'll make it easier to make
with.

A tab structure has been setup, where one tab has all the metadata info
and the other tab has the files. The files tab now has two display
options, one is a tiled view and the other is a list view. I didn't have
enough time to do the list view justice, so currently, it's just the
previous file listing code. One issue with the tile view is that file
descriptions aren't displayed. The file description is in the tooltip on
hover over the media/thumbnail area. The problem is that since file
descriptions can be varied in length or non-existant, it'd kill the
equal height boxes in the tile view. It's such a pain to make equal
height boxes in CSS, in the future, if we have more time, I'd like to
see Bootstrap v3 replaced with something newer.

The sidebar has been subsumed into the main body, as it really was
taking up too much space.

The Admin Tools sidebar has also been refactored into a dismissable
alert box at the top of the page.

Implements several tasks specified in #48.