EDIT Attempt to overhaul edit Date field.
The fields generated by edit-metadata.jsp is very difficult to modify
due to them being generated by an unholy mix of scriptlets and html. Was
trying to implement some kind of date validation for the date field but
decided to try overhauling it for easier modification in the future. The
attempt was not really successful. There was too much logic in the
scriptlets to completely port in a day. One issue is retrieval of data
on repeatable elements. Due to limitations of JSP includes, I can't pass
in arrays of data unfortunately. So I've resorted to falling back to the
old field generation code for repeatable date fields, only using the new
code for single date fields. For StatSpace, this is sufficient as we
don't have any repeatable date fields.

Next up is again try to validate dates with the new easier to modify
field generation code.