EDIT Modernize the date input.
Using the jQuery UI datepicker. This modified the date component
introduced previously. It ditched pretty much all the old input fields
for the jQuery UI one.

Note that the original complaint that this is addressing is the input of
invalid dates. Given the lack of time, I couldn't investigate adding the
validation on the server side. The jQuery UI datepicker just makes the
date selection easier, which kinda sidesteps users inputting invalid
dates since they're picking from choices that we select. It doesn't
actually prevent bad input if the user is hell bent on it.

Also fixed some nagging formatting issues where the input boxes on the
submission form wasn't lined up.