EDIT media streaming now checks browser compatible
Streaming media will check mimetype with browser to see if they can play
the content. This isn't perfect as our mime type doesn't include codec
information. If not compatible, will display an error message.

Also moved the video/audio tags into components as they were being used
in two different places.

Updated videojs to latest 7.0.3.

For some reason, the header height was broken in Chrome. Can't figure
out why, so put a rough patch on it for now.

Found out that audio/wave DOES NOT work in Chrome. audio/wav works in
both Chrome and Firefox, so changed the wave mimetype to that.