FIX special char in filenames break download.
Noticed that when you download filenames with spaces in them, you end up
with a file that has a truncated name, e.g.: filename "test 123.csv" you
get a file named "test".

Looks like the HTML5 solution is to use the download attribute.
Unfortunately, this isn't supported in IE11. So for IE11 and below,
adding the filename to the Content-Disposition header is the fix.

Firefox requires that filenames be quoted in the Content-Disposition
header, otherwise, filenames will be truncated if there's a space or
other special characters.

Looks like there was special treatment for large files where they're
force downloaded. Large files will have the "Content-Disposition:
attachment" header set. It doesn't seem to interfere with partial
content requests weirdly enough. I've combined the forcedownload flag
with this mechniasm so that it's all in one place.

Fixes #45.