FIX Previous button in submission kills submission
Clicking "Previous" from "Upload" stage to go back to the "Describe"
stage throws an "Internal Server Error". Checking the logs, it seems
that parameters aren't passed properly to date.jsp. The request
attributes are registered properly when used directly on page, but when
passed to jsp:include, it doesn't work. Not sure why.

Workaround is to just put the date stuff directly on page for now. No
time to troubleshoot further.

This is task number 52 in the task list #51.