EDIT defaults to 4 fields for repeatable oneboxes.
As a workaround for clicking "Add More" shifts the user view,
pregenerate more fields to reduce the need for the user to click "Add
More".

Note that the original request was to limit to prereqs, learning
objectives, and related materials. But applying this to every repeatable
onebox for now and will revisit if I have time.

This is task number 56 in task list #51.