EDIT user vetting now opt-in.
Previous behaviour was that ALL registered users get vetted. The desired
behaviour is that users can, at the time of registration or by editing
their profile, tick a checkbox to request to be vetted.

Note that this commit requires an manual update to the metadata schema
for eperson, you need to add the following fields:
- eperson.additionalInfo
- eperson.institution.address
- eperson.institution.type
- eperson.institution.url
- eperson.requestedInstructorAccess
- eperson.supervisor

This is take number 13 in task list #51. This finishes #51.