EDIT license select, validate only if next step.
The license selection page requires users to check the license agreement
boxes to continue. This unfortunately also affects the other buttons
such as "Previous" "Cancel" and jumping to other previously completed
stages of the submission process. So the user can't backtrack without
also checking those checkboxes.

Since we're using the html5 required attribute for this, we can disable
the form validation by putting the noformvalidate attribute on the
submit buttons.

This is task number 93 on task list #52.