FIX hard coded demo JSPs throwing errors.
For the about and copyright JSPs: The pages weren't found because they
were in the wrong place. Forgot to move them. Weird error it throws when
a JSP isn't found, it says "Cannot forward after response has been
committed".

For the other JSPs, they were throwing errors even though they rendered
fine. This turns out to be because I was calling showJSPs consecutively.
I was mistakenly treating it like a return instead of a function call.