FIX restricted file indicator, hidden controls.
On the upload file page, the submission was noted as restricted even
though "Open" restrictions was set. This is due to a mismatch between
the restriction constants. One was configured to "Everyone" while the
other was configured to "Open". This is due to input-forms.xml having a
version of the constant which has to be duplicated in
Messages.properties. Just needed to match them up. "Open" is was
CWSEISpace wants for submissions that can be accessed by everyone.

Some controls, notably in the selecting users or groups, were hidden due
to improper usage of bootstrap classes. There really should only be one
element with the "container" class. It seems that the usage works fine
in older versions of bootstrap, but since we updated bootstrap, they no
longer behave as they used to, elements with the "container" class
doesn't obey their parent's size restrictions, so sometimes, they extend
offscreen hiding controls.

The discovery pages suffer from this and also the fact that the
"pull-right" class means that the element's height isn't taken into
account when calculating the parent's height. This means that the page
number links collide with the search results.