ADD ability to assign forms to Communities.
In stock DSpace, the forms configured in input-forms.xml can only be
assigned to the collection level. This is inconvenient for CWSEISpace
since organizationally, a Community is a subject and they have subject
specific forms. So I've built in the ability to accept community-name
attributes in the form mapping.

Note that community names that doesn't have a match with existing
communities will just be ignored. An error is generated in the log but
execution otherwise continues.

Since handles are instance specific, I had to keep a different
input-forms.xml in my repo for my testing instance. Switching to
community names means that I no longer have to do that, just need to
have communities with the same names as CWSEISpace is using.