EDIT use real servlet for serving home page.
index.jsp was apparently being used as a ghetto home page servlet. Looks
like before Tomcat7, there was issues with the web.xml not being able to
specify servlets for the index, I guess this was their workaround. Since
we don't have to deal with that limitation, I've moved the relevant code
back into a home servlet. This goes toward explaining the mystery of how
the menu bar on the home page can recognize that users are logged in but
wouldn't register it on the rest of the page.