FIX navbar home links & logged in home page.
The logged in home page was showing original DSpace content since the
user data is retrivable now with the home page servlet active. This
isn't intended as Stats want the hard coded home page.

Links to the home page could technically still use /home.jsp but that
bypasses the servlet and so might produce weird behaviour. I've fixed
this for all the navbars. The hard coded content still links to
/home.jsp, but I've no plans to fix those as they should be temporary
anyways.