EDIT home page mockup, round 2 refinements.
A rearrangement and compacting of the original mockup. The subjects list
are now clickable links with each going to a search page looking for
that particular subject.

The featured articles carousel is powered by the slick library. Slick is
a jquery based carousel library. Bootstrap does have a carousel library,
but it depends on an image to establish carousel height and also doesn't
seem to cope well when shrunk to mobile view. Note that I've modified
the slick css theme slightly in order to make it compatible with the
white background on BioSpace.

Subject icons from flaticon, license files will be documented in gitlab.

This completes task 8,9,23 in #55 and #56.