EDIT submission fields ordering, hints, removal.
All required fields are now shown first to the user, followed by
optional fields.

Summary/Abstract is now short description. A one or two sentence
description of the item. Note that the long form description was removed
as well, so might have to an optional field for a longer description if
people wants to enter those.

Language now an optional field. Removed "available in alternate
language".

Pre-req fields now reduced to 2 entries from 4 by default. Hints were
updated to be relevant to biology.

This completes tasks 10,11,12,13,14 in #55.