ADD "download all files" tab in display item.
When looking at items, a user might want to download all the files
attached to the item in one go instead of having to download them
individually. The "Download All Files" tab will let users download all
the files as a zip package. This requires each item to have a zip
package generated. The generation is done by a Quartz job that triggers
every 30 minutes. Two files are actually generated, one a restricted
version and one an unrestricted version. The restricted version contains
instructor-only files while the unrestricted one doesn't.

A servlet was needed to serve the appropriately permissioned file to
useres.

While DSpace stock has an ability to export items as zip files, it seems
to be restricted to administrators only. The export is also
user-specific and meant for import to other DSpace instances. I've used
their export code as a basis for the package file generation here. Which
means that package files are generated in the same directory as the
stock exports. Except that instead of creating subdirectories based on
user ID, I've put them in a the "all" directory.

This is task 18 in #55. Requested by BioSpace, but StatSpace could use
it too.