EDIT datepicker allow separate year/month/date.
Stock DSpace only requires year to be entered. When the datepicker was
implemented, the datepicker wasn't flexible enough to duplicate this.
You needed to pick a specific date with the entire all of
year/month/date.

Bio expressed a desire for only picking year or year and month, etc. So
I decided to tackle this again. Found another datepicker that was more
flexible. This one allows picking year/month/date in sequence, so
partial entries are possible. Downside is that it does take longer to
pick up how to use it. Bio also pointed out that users might just enter
the dates manually, so I've allowed for that possibilty too with
something like stock DSpace's separated year-month-date fields.

This is task 26 for task list #56.