EDIT separate datepicker for each year/month/date.
Instead of using one datepicker for picking all 3 year/month/date,
there's not a separate instance on each of those fields. I figure this
would be more intuitive for users.

This turned out to be more difficult than I thought due to the need to
keep all 3 datepicker instances synced up to the same date. It wouldn't
make much sense if changing the month in the month field shows you a
different month in the date field.

This is task 26 in task list #56.