EDIT subject multiselect, hide accesss restriction
The subject field in submission form is now a multi-select.

"Access Restrictions" is no longer configurable and will default to
"Everyone" access. This necessitates the configuration of default
metadata for dcterms.accessRights documented in #34.

Also decreased speed of featured articles carousel.

This is task 24, 29, and partial 27 in task lists #56 and #57.