EDIT de-emphasize resource type in search results.
UI review noted that it was too heavy visually, which doesn't make sense
for their purpose.

Also tried to make them look similar to the mockup featured articles
panels on the front page. The problem is that the mockup panels have a
lot more space to work with, so I'm not sure about showing the resource
types in normal sized font. Have put them inside a <small> tag for now.