FIX thumbnail only looking at first file.
If the first file doesn't have a thumbnail, then the entire resource
doesn't have a thumbnail. Even if other files in the resource has a
thumbnail. This is because it was only looking at the first file in each
resource for a thumbnail. Fixed it so that it'll go through each file in
order to find a thumbnail for the resource.