FIX UI consistent use of Join, navbar highlight.
Fixes from the UI review.

Navbar will highlight "Join" and "Sign in" button when on those
respective pages.

Consistently use "Join" instead of "Register" on the user registration
page.

Added an icon for the "Join" button for consistency with other buttons
that has an icon. At least for the non-logged in user.

Removed "Join" section on the copyright page.