EDIT migrate to new rewritten edit-metadata.jsp.
Moved StatSpace specific stuff over to the new edit-metadata.jsp that
was rewritten from the grounds up, see CWSEISpace commit for details.

The main thing to migrate was the triple level subjects dropdown.
Instead of patching DCInput for new input types, I've modified FieldInfo
to check for the triple level subjects dropdown instead. So
edit-metadata.jsp will handle it as a new input type.

Note that while migrating, I've removed the code that allows repeatable
subjects as we aren't using it.

Also implemented the UI review recommendation for setting the Subject
selection to a blank selection initially, with instructions that make it
clear that there's hierarchical structure behind the selections.

Show/Hide Subject (Other) was taken from BioSpace's version and
modified. See BioSpace's commit for details. I've removed the message
asking users to enter a custom subject as the show Subject (Other)
animation highlights and field and hopefully should be enough to draw
the user's attention.