EDIT search results will only return items.
This is a long standing issue where the search results will show you
communities and collections. I've been working around it by not showing
communities and collections in the results, but the search result count
& pagination still takes those hidden communities and collections into
account. So we end up with cases where there's a full page of empty
results.

Turns out there's a way to configure the discovery layer to only find
items. The commented out feature is only in one location, but
uncommenting it turns out to be insufficient. Turns out you have to
duplicate it in another section in the discovery configuration for it to
take effect.

For reference:
http://dspace.2283337.n4.nabble.com/Re-Communities-and-Collections-in-search-results-td4672768.html