EDIT complete rewrite of search.
This is an initial implementation of the search improvements that
Stat/Bio wants. As the stock search page was severely lacking in good
UI and the implementation itself is state-of-the-art early 2000s, a
lengthy rewrite was unfortunately required to make the improvements
feasible.

The template for layout and behaviour was UBC library circle database,
which also uses a heavily customized version of DSpace:
https://open.library.ubc.ca/search

E.g.: The Discovery facets was moved to a left sidebar and is collapsed
by default. I've also gotten rid of the facet's default paging
behaviour. The search results now default into a list. Search filters
are now hidden behind the Advanced Search button.

I've maintained the old tile view as a toggleable option. Hence why the
default results per page is still multiples of 9. Might revisit later on
to see if we can change it to back to even numbers.

Some features haven't been implemented and some bugs remain, notably:
autocomplete for filter queries, persist list/tile view selection, low
resolution UBC logo placeholder.

The code could also use some cleanup, remove item javascript was copied
from edit-metadata.jsp and modified slightly. It's probably better to
generalize the two use cases into one code base so we don't have
duplicate code. Also, a lot of unused attributes could probably be
trimmed in the UBCDiscoverySearchRequestProcessor.

Note that the result filters use bootstrap's collapse js, I'm not sure
how performant it is. I've been using jquery-ui for show/hide, not sure
if we should all use jquery-ui for better performance.

This implements task 1,2,3,4 in task list #59.